Ep 07 - Taking Networked damages

+++ Assets/Scripts/PlayerWeapon.cs
@@ -4,6 +4,6 @@
 public class PlayerWeapon
 {
     public string name = "Glock";
-    public float damage = 10f;
+    public int damage = 10;
     public float range = 100f;
 }
